fix: only call NetworkService::SetUpHttpAuth once.
